Editor: Unify text/code editor between post and site editors
