+++ packages/editor/src/components/visual-editor/edit-template-blocks-notification.js
