+++ packages/editor/src/components/visual-editor/style.scss
@@ -1,4 +1,4 @@
-.editor-canvas {
+.editor-visual-editor {
 	height: 100%;
 
 	&.is-resizable {
