+++ packages/editor/src/private-apis.js
@@ -6,7 +6,6 @@ import * as interfaceApis from '@wordpress/interface';
 /**
  * Internal dependencies
  */
-import EditorCanvas from './components/editor-canvas';
 import { ExperimentalEditorProvider } from './components/provider';
 import { lock } from './lock-unlock';
 import { EntitiesSavedStatesExtensible } from './components/entities-saved-states';
@@ -31,6 +30,8 @@ import PostContentInformation from './components/post-content-information';
 import PostLastEditedPanel from './components/post-last-edited-panel';
 import ResizableEditor from './components/resizable-editor';
 import Sidebar from './components/sidebar';
+import TextEditor from './components/text-editor';
+import VisualEditor from './components/visual-editor';
 import {
 	mergeBaseAndUserConfigs,
 	GlobalStylesProvider,
@@ -40,7 +41,6 @@ const { store: interfaceStore, ...remainingInterfaceApis } = interfaceApis;
 
 export const privateApis = {};
 lock( privateApis, {
-	EditorCanvas,
 	ExperimentalEditorProvider,
 	EntitiesSavedStatesExtensible,
 	GlobalStylesProvider,
@@ -64,6 +64,8 @@ lock( privateApis, {
 	PostLastEditedPanel,
 	ResizableEditor,
 	Sidebar,
+	TextEditor,
+	VisualEditor,
 
 	// This is a temporary private API while we're updating the site editor to use EditorProvider.
 	useAutoSwitchEditorSidebars,
