+++ packages/editor/src/style.scss
@@ -8,7 +8,6 @@
 @import "./components/document-bar/style.scss";
 @import "./components/document-outline/style.scss";
 @import "./components/document-tools/style.scss";
-@import "./components/editor-canvas/style.scss";
 @import "./components/editor-notices/style.scss";
 @import "./components/entities-saved-states/style.scss";
 @import "./components/error-boundary/style.scss";
@@ -52,3 +51,5 @@
 @import "./components/site-discussion/style.scss";
 @import "./components/table-of-contents/style.scss";
 @import "./components/template-areas/style.scss";
+@import "./components/text-editor/style.scss";
+@import "./components/visual-editor/style.scss";
